음악 API의 DB 부분을 거의 완성하였습니다. 이제 파일을 저장하는 로직을 구현해야합니다.

+++ src/main/java/com/binn328/ym_player/Controller/MusicAPIController.java
@@ -1,10 +1,11 @@
 package com.binn328.ym_player.Controller;
 
-import com.binn328.ym_player.DAO.Music;
+import com.binn328.ym_player.Model.Music;
 import com.binn328.ym_player.DTO.MusicForm;
 import com.binn328.ym_player.Repository.MusicRepository;
-import com.binn328.ym_player.Service.MusicService;
 import com.binn328.ym_player.Service.StorageService;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -12,6 +13,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 import java.util.Optional;
 
+@Log4j2
 @RestController
 @RequestMapping("/api/music")
 public class MusicAPIController {
@@ -48,21 +50,69 @@ public class MusicAPIController {
 
     /**
      * 음악을 추가합니다.
+     * 우선 데이터베이스에 저장하여 id를 가져온 후, 해당 id를 이름으로 파일을 저장합니다.
      * @param form 추가될 음악의 정보입니다.
      * @return
      */
-    // https://youtu.be/mi5eKp7noRw?list=PLfu_Bpi_zcDPtiIMdKbQ33OnqzAf9lRjg&t=997
-    // api 만들고, @RequestPart form 을 사용하도록 리팩토링하기
-    @PostMapping(consumes = "")
+    @PostMapping(consumes = "multipart/form-data")
     public ResponseEntity<Music> addMusic(MusicForm form, @RequestPart("file") MultipartFile file) {
-        // TODO 저장기능 추가
+
         Music music = form.toEntity();
-        musicRepository.save(music);
-        return null;
+        // 가져온 음악 정보 로깅
+        log.info(music.toString());
+        Music savedMusic = musicRepository.save(music);
+        storageService.saveMusic(file, savedMusic.getId());
+
+        return new ResponseEntity<>(savedMusic, HttpStatus.CREATED);
     }
 
-    @DeleteMapping("api/music")
-    public void deleteMusic(String id) {
-        musicRepository.deleteById(id);
+    @PutMapping("/{id}")
+    public ResponseEntity<Music> updateMusic(MusicForm form, @PathVariable String id) {
+        // TODO JSON 을 받게 수정하거나, musicForm 대신 그냥 Music을 넣어도 되지 않을까?
+        Music musicUpdateData = form.toEntity();
+        Optional<Music> musicOptional = musicRepository.findById(id);
+        if (musicOptional.isPresent()) {
+            Music musicToUpdate = musicOptional.get();
+            // 노래 제목 업데이트
+            if (musicUpdateData.getTitle() != null) {
+                musicToUpdate.setTitle(musicUpdateData.getTitle());
+            }
+            // 가수 업데이트
+            if (musicUpdateData.getArtist() != null) {
+                musicToUpdate.setArtist(musicUpdateData.getArtist());
+            }
+            // 그룹 업데이트
+            if (musicUpdateData.getGroup() != null) {
+                musicToUpdate.setGroup(musicUpdateData.getGroup());
+            }
+            // 좋아요 여부 업데이트
+            musicToUpdate.setFavorite(musicToUpdate.isFavorite());
+
+            // db에 저장
+            musicRepository.save(musicToUpdate);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+    /**
+     * 음악을 제거합니다.
+     * @param id 제거될 음악의 id 입니다.
+     */
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Music> deleteMusic(@PathVariable String id) {
+        // TODO 음악이 제대로 제거되었는지 확인하기
+
+        if (musicRepository.existsById(id)) {
+            storageService.deleteMusicById(id);
+            musicRepository.deleteById(id);
+
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+
+
     }
 }
